backend/requirements.txt
flask
alembic
flask-sqlalchemy
celery
flower
Flask-Mail
MySQL-python
feedparser
redis
pytz>=2013b
requests
lxml
flask-script
flask-restful
flask-assets
flask-httpauth
passlib
py-bcrypt
flask-seasurf
cssmin
fabric
python-dateutil
beautifulsoup4
supervisor
uwsgi